Refactor functions to improve readability
