+++ src/utils/walking.test.ts
@@ -1,0 +1,3 @@
+import { describe, expect, it } from 'vitest';
+
+describe.skip('walker', () => {});
